Refactor item model imports to use common item model and update related components
