+++ src/components/home-components/featured-items.component.tsx
@@ -1,9 +1,9 @@
 import React, { Suspense } from "react";
 import DefaultLoaderComponent from "../ui/default-loader.component";
-import { ItemModel } from "@/models/product-category-collections/item.model";
 import { itemAPI } from "@/services/item.service";
 import ProductCardComponent from "../cards/product-card.component";
 import { getAuthToken } from "@/utils/cookie-store";
+import { ItemModel } from "@/models/common/item.model";
 
 async function FeaturedItemComponent() {
   const token = await getAuthToken();
