+++ src/components/home-components/new-arrival.component.tsx
@@ -1,9 +1,9 @@
-import { ItemModel } from "@/models/product-category-collections/item.model";
 import { itemAPI } from "@/services/item.service";
 import React, { Suspense } from "react";
 import ProductCardComponent from "../cards/product-card.component";
 import DefaultLoaderComponent from "../ui/default-loader.component";
 import { getAuthToken } from "@/utils/cookie-store";
+import { ItemModel } from "@/models/common/item.model";
 
 async function NewArrivalComponent() {
   const token = await getAuthToken();
