+++ src/components/product-detail-components/item-details-and-pricing.component.tsx
@@ -1,4 +1,4 @@
-import { ItemModel } from "@/models/product-category-collections/item.model";
+import { ItemModel } from "@/models/common/item.model";
 import firstLetterCapitalize from "@/utils/first-letter-capitalize";
 import priceFormatter from "@/utils/price-formatter";
 import React from "react";
